fix autirization hook and added page Start

+++ client/src/hooks/useAuth.ts
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { LOGIN_ROUTE } from '@utils/contants';
+import { LOGIN_ROUTE, START_ROUTE } from '@utils/contants';
 
 import { AUTH_HEADER } from '@src/utility/headers';
 import { $api } from '@src/http/http';
@@ -12,13 +12,17 @@ const storageService = StorageService.getInstance();
 
 const useAuth = () => {
   const token = storageService.getItem(AUTH_HEADER);
-  console.log(token);
+  const isPreviewChecked = storageService.getItem('previewCheck');
   const navigate = useNavigate();
 
   useEffect(() => {
+    if (token === null && !isPreviewChecked) {
+      navigate(START_ROUTE);
+    }
     if (token !== null) {
       $api.defaults.headers.common[AUTH_HEADER] = token;
-    } else {
+    }
+    if (token === null && isPreviewChecked) {
       storageService.removeItem(AUTH_HEADER);
       storageService.removeItem('activeTabLC');
       navigate(LOGIN_ROUTE);
